+++ shuup/core/models/_product_media.py
@@ -36,27 +36,47 @@ class ProductMediaKind(Enum):
 class ProductMedia(TranslatableModel):
     identifier = InternalIdentifierField(unique=True)
     product = models.ForeignKey("Product", related_name="media", on_delete=models.CASCADE, verbose_name=_('product'))
-    shops = models.ManyToManyField("Shop", related_name="product_media", verbose_name=_('shops'))
+    shops = models.ManyToManyField("Shop", related_name="product_media", verbose_name=_('shops'), help_text=_(
+            "Select which shops you would like the product media to be visible in."
+        )
+    )
     kind = EnumIntegerField(
-        ProductMediaKind, db_index=True, default=ProductMediaKind.GENERIC_FILE, verbose_name=_('kind')
+        ProductMediaKind, db_index=True, default=ProductMediaKind.GENERIC_FILE, verbose_name=_('kind'), help_text=_(
+            "Select what type the media is. It can either be a normal file, part of the documentation, or a sample."
+        )
     )
     file = FilerFileField(blank=True, null=True, verbose_name=_('file'), on_delete=models.CASCADE)
     external_url = models.URLField(
         blank=True, null=True, verbose_name=_('URL'),
         help_text=_("Enter URL to external file. If this field is filled, the selected media doesn't apply.")
     )
-    ordering = models.IntegerField(default=0, verbose_name=_('ordering'))
+    ordering = models.IntegerField(default=0, verbose_name=_('ordering'), help_text=_(
+            "You enter the numerical order that your image will be displayed on your product page."
+        )
+    )
 
     # Status
     enabled = models.BooleanField(db_index=True, default=True, verbose_name=_("enabled"))
-    public = models.BooleanField(default=True, blank=True, verbose_name=_('public (shown on product page)'))
+    public = models.BooleanField(
+        default=True, blank=True, verbose_name=_('public (shown on product page)'), help_text=_(
+            "Check this if you would like the image shown on your product page. Checked by default."
+        )
+    )
     purchased = models.BooleanField(
-        default=False, blank=True, verbose_name=_('purchased (shown for finished purchases)')
+        default=False, blank=True, verbose_name=_('purchased (shown for finished purchases)'), help_text=_(
+            "Select this if you would like the product media shown for completed purchases."
+        )
     )
 
     translations = TranslatedFields(
-        title=models.CharField(blank=True, max_length=128, verbose_name=_('title')),
-        description=models.TextField(blank=True, verbose_name=_('description')),
+        title=models.CharField(blank=True, max_length=128, verbose_name=_('title'), help_text=_(
+                "Choose a title for your product media. This will help it be found in your store and on the web."
+            )
+        ),
+        description=models.TextField(blank=True, verbose_name=_('description'), help_text=_(
+                "Write a description for your product media. This will help it be found in your store and on the web."
+            )
+        ),
     )
 
     class Meta: